requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
httpx>=0.27.0
python-dotenv>=1.0.0
eth-account>=0.13.0
eth-utils>=5.0.0
cryptography>=42.0.0
base58>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
x402[httpx,evm]>=2.6.0
